+++ src/telegram/main.js
@@ -4,6 +4,7 @@ const { getDatabase, onValue, ref } = require("firebase/database");
 const { START_MESSAGE } = require("./constants");
 const { TRIGGERS_DB } = require("../bot/constants");
 const { getUserId } = require("../bot/database");
+const { updateUser } = require("../admin-panel/database");
 const {
 	getGroups,
 	addGroup,
@@ -12,6 +13,10 @@ const {
 	getServices,
 	addService,
 	resetServices,
+	addOngoingService,
+	removeOngoingService,
+	getServedUser,
+	checkOngoingService,
 } = require("./database");
 
 const bot = new Telegraf(process.env.TELEGRAM_TOKEN);
@@ -82,12 +87,51 @@ bot.command("reset_leader_board", async (ctx) => {
 	}
 });
 
+bot.command("finish", async (ctx) => {
+	const description = ctx.update.message.text.replace("/finish", "").trim();
+	const username = ctx.from.username;
+	if (!await checkOngoingService(database, username)) {
+		await ctx.reply("You don't have ongoing services 🧙");
+		return;
+	}
+	const userId = await getServedUser(database, username);
+	if (description.length > 0) {
+		await updateUser(userId, description, database);
+	}
+	Promise.all([
+		removeOngoingService(database, username),
+		addService(database, username, userId),
+	]);
+	await ctx.reply("Congratulations on successful service! 🎉");
+});
+
 bot.action(/pick:(.+)/, async (ctx) => {
+	const username = ctx.from.username;
+	if (await checkOngoingService(database, username)) {
+		console.log("How am I here?");
+		await ctx.answerCbQuery("You already have an ongoing service");
+		return;
+	}
 	const userId = ctx.match[1];
 	ctx.answerCbQuery();
-	const username = ctx.from.username;
-	await addService(database, username, userId);
-	await ctx.editMessageText(`@${username} picked ${userId} 🤩`);
+	const message = `@${username} 🥞\nYou picked a client with ID +${userId}. ⛳️\nTo complete service and receive points, \n1) Go to https://wa.me/+${userId} or call +${userId} 🍄\n2) Once you're done, click below and follow instructions ⬇️⬇️⬇️`;
+	const keyboard = Markup.inlineKeyboard([
+		Markup.button.callback("Done ✅", `done:${userId}:${username}`),
+	]);
+	await addOngoingService(database, username, userId);
+	await ctx.editMessageText(message, keyboard);
+});
+
+bot.action(/done:(.+):(.+)/, async (ctx) => {
+	const userId = ctx.match[1];
+	const username = ctx.match[2];
+	if (!checkOngoingService(database, username, userId)) {
+		ctx.answerCbQuery("You don't serve this client");
+		return;
+	}
+	ctx.answerCbQuery();
+	const message = `@${username} We are almost done! 😺\n To finish serving this client, type /finish 🤠\nYou can optionally specify some additional information on this user right after command 👥`
+	await ctx.editMessageText(message)
 });
 
 const dbRef = ref(database, TRIGGERS_DB);